--- client/src/pages/AccountPage.jsx
@@ -1,153 +1,0 @@
-import { useContext, useState } from "react";
-import { UserContext } from "../UserContext";
-import { Link, Navigate, useParams } from "react-router-dom";
-import axios from "axios";
-
-//import StaysPage from "./StaysPage";
-
-import PlacesPage from "./PlacesPage";
-
-//*************Account Page function with export included */
-export default function AccountPage() {
-  const [redirect, setRedirect] = useState(null);
-
-  const { ready, user, setUser } = useContext(UserContext);
-
-  let { subpage } = useParams();
-  if (subpage === undefined) {
-    subpage = "profile";
-  }
-
-  async function logout() {
-    await axios.post("/logout");
-    setRedirect("/");
-    setUser(null);
-  }
-
-  //If user is not ready console out loading
-  if (!ready) {
-    return "loading...";
-  }
-
-  //if ready and there is no user go to login
-  if (ready && !user && !redirect) {
-    return <Navigate to={"/login"} />;
-  }
-
-  function LinkClasses(type = null) {
-    let classes = "inline-flex gap-1 py-2 px-6 rounded-full";
-    if (type === subpage || (subpage === undefined && type === "profile")) {
-      classes += " bg-primary rounded-full"; // Concatenate the classes using +=
-    } else {
-      classes += "bg-gray-200";
-    }
-    return classes;
-  }
-
-  if (redirect) {
-    return <Navigate to={redirect} />;
-  }
-
-  return (
-    <div>
-      <nav className="w-full flex justify-center mt-8 gap-2 mb-8">
-        <Link className={LinkClasses("profile")} to={"/account"}>
-          <svg
-            xmlns="http://www.w3.org/2000/svg"
-            fill="none"
-            viewBox="0 0 24 24"
-            strokeWidth={1.5}
-            stroke="currentColor"
-            className="w-6 h-6"
-          >
-            <path
-              strokeLinecap="round"
-              strokeLinejoin="round"
-              d="M17.982 18.725A7.488 7.488 0 0012 15.75a7.488 7.488 0 00-5.982 2.975m11.963 0a9 9 0 10-11.963 0m11.963 0A8.966 8.966 0 0112 21a8.966 8.966 0 01-5.982-2.275M15 9.75a3 3 0 11-6 0 3 3 0 016 0z"
-            />
-          </svg>
-          My Profile
-        </Link>
-        <Link className={LinkClasses("bookings")} to={"/account/bookings"}>
-          <svg
-            xmlns="http://www.w3.org/2000/svg"
-            fill="none"
-            viewBox="0 0 24 24"
-            strokeWidth={1.5}
-            stroke="currentColor"
-            className="w-6 h-6"
-          >
-            <path
-              strokeLinecap="round"
-              strokeLinejoin="round"
-              d="M12.75 3.03v.568c0 .334.148.65.405.864l1.068.89c.442.369.535 1.01.216 1.49l-.51.766a2.25 2.25 0 01-1.161.886l-.143.048a1.107 1.107 0 00-.57 1.664c.369.555.169 1.307-.427 1.605L9 13.125l.423 1.059a.956.956 0 01-1.652.928l-.679-.906a1.125 1.125 0 00-1.906.172L4.5 15.75l-.612.153M12.75 3.031a9 9 0 00-8.862 12.872M12.75 3.031a9 9 0 016.69 14.036m0 0l-.177-.529A2.25 2.25 0 0017.128 15H16.5l-.324-.324a1.453 1.453 0 00-2.328.377l-.036.073a1.586 1.586 0 01-.982.816l-.99.282c-.55.157-.894.702-.8 1.267l.073.438c.08.474.49.821.97.821.846 0 1.598.542 1.865 1.345l.215.643m5.276-3.67a9.012 9.012 0 01-5.276 3.67m0 0a9 9 0 01-10.275-4.835M15.75 9c0 .896-.393 1.7-1.016 2.25"
-            />
-          </svg>
-          My Bookings
-        </Link>
-        <Link className={LinkClasses("places")} to={"/account/places"}>
-          <svg
-            xmlns="http://www.w3.org/2000/svg"
-            fill="none"
-            viewBox="0 0 24 24"
-            strokeWidth={1.5}
-            stroke="currentColor"
-            className="w-6 h-6"
-          >
-            <path
-              strokeLinecap="round"
-              strokeLinejoin="round"
-              d="M8.25 6.75h12M8.25 12h12m-12 5.25h12M3.75 6.75h.007v.008H3.75V6.75zm.375 0a.375.375 0 11-.75 0 .375.375 0 01.75 0zM3.75 12h.007v.008H3.75V12zm.375 0a.375.375 0 11-.75 0 .375.375 0 01.75 0zm-.375 5.25h.007v.008H3.75v-.008zm.375 0a.375.375 0 11-.75 0 .375.375 0 01.75 0z"
-            />
-          </svg>
-          My Accommodations
-        </Link>
-      </nav>
-
-      {subpage === "profile" && (
-        <div className="text-center max-w-lg mx-auto ">
-          Logged in as {(user.name, user.email)}
-          <br />
-          <button onClick={logout} className="primary max-w-sm ">
-            Log out
-          </button>
-        </div>
-      )}
-      {subpage === "places" && <PlacesPage />}
-    </div>
-  );
-}
-
-// export default function AccountPage() {
-//   const [redirect, setRedirect] = useState(null);
-//   const { ready, user, setUser } = useContext(UserContext); //use useCOntext fro UserCOntext to get status of user and ready
-
-//   async function logout() {
-//     await axios.post("/logout");
-//     setRedirect("/");
-//     setUser(null);
-//   }
-
-//   if (!ready) {
-//     return "Loading....";
-//   }
-
-//   //If statements to handle potential null or undefined values:
-//   //If not ready and theres no user or no user name then navigate to login page
-//   if ((ready && !user) || !user.name) return <Navigate to={"/login"} />;
-// //   //function to highlight link of current page.
-//   function LinkClasses(type = null) {
-//     let classes = "py-2 px-6";
-//     if (type === subpage) {
-//       classes += " bg-primary rounded-full"; // Concatenate the classes using +=
-//     }
-//     return classes;
-//   }
-
-//   //**********Button Links STAYS AND LISTINGS**********/
-
-//   if (redirect) {
-//     return <Navigate to={redirect} />;
-//   }
-
-// }
